+++ CHANGELOG.md
@@ -4,6 +4,22 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.8.0] - 2020-02-23
+
+### Added
+- Inject a `application/octet-stream` as default content type when none is
+  present (#140, thanks requaos)
+- Add support for content-type params to part & encoding (#148, thanks
+  pzeinlinger)
+- UTF-7 support (#17)
+
+### Fixed
+- Handle missing parameter values in the middle of the media parameter list
+  (#139, thanks requaos)
+- Fix boundaryReader to respect length instead of capacity (#145, thanks
+  dcormier)
+- Handle very empty mime parts (#144, thanks dcormier)
+
 ## [0.7.0] - 2019-11-24
 
 ### Added
@@ -114,7 +130,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Initial implementation of MIME encoding, using `enmime.MailBuilder`
 
 [Unreleased]: https://github.com/jhillyerd/enmime/compare/master...develop
-[0.6.0]:      https://github.com/jhillyerd/enmime/compare/v0.6.0...v0.7.0
+[0.8.0]:      https://github.com/jhillyerd/enmime/compare/v0.7.0...v0.8.0
+[0.7.0]:      https://github.com/jhillyerd/enmime/compare/v0.6.0...v0.7.0
 [0.6.0]:      https://github.com/jhillyerd/enmime/compare/v0.5.0...v0.6.0
 [0.5.0]:      https://github.com/jhillyerd/enmime/compare/v0.4.0...v0.5.0
 [0.4.0]:      https://github.com/jhillyerd/enmime/compare/v0.3.0...v0.4.0
@@ -133,5 +150,5 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 3.  Run tests
 4.  Commit changes and merge release: `git flow release finish`
 
-See http://keepachangelog.com/ for additional instructions on how to update this
+See http://keep change log.com/ for additional instructions on how to update this
 file.